requirements.txt
asgiref==3.8.1
Django==5.0.6
Markdown==3.6
sqlparse==0.5.0
pymysql==1.1.1
line-bot-sdk==3.11.0